requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.14.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
attrs==21.2.0
    # via jsonschema
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.6
    # via requests
clang==5.0
    # via tensorflow
flatbuffers==1.12
    # via tensorflow
gast==0.4.0
    # via tensorflow
google-auth==1.35.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.41.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.1.0
    # via tensorflow
idna==3.2
    # via requests
jsonschema==3.2.0
    # via valohai-yaml
keras==2.6.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.4
    # via tensorboard
numpy==1.19.5
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.1.1
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
pillow==8.3.2
    # via -r requirements.in
protobuf==3.18.0
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyrsistent==0.18.0
    # via jsonschema
pyyaml==5.4.1
    # via valohai-yaml
requests==2.26.0
    # via
    #   requests-oauthlib
    #   tensorboard
    #   valohai-utils
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   jsonschema
    #   keras-preprocessing
    #   tensorflow
tensorboard==2.6.0
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorflow==2.6.0
    # via -r requirements.in
tensorflow-estimator==2.6.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
tqdm==4.62.3
    # via valohai-utils
typing-extensions==3.7.4.3
    # via tensorflow
urllib3==1.26.7
    # via requests
valohai-papi==0.1.3
    # via valohai-utils
valohai-utils==0.1.11
    # via -r requirements.in
valohai-yaml==0.15.0
    # via
    #   valohai-papi
    #   valohai-utils
werkzeug==2.0.1
    # via tensorboard
wheel==0.37.0
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools